Add the possibility to start the server as a daemon with zogShell.
The commands are:
- server.daemonize
- server.exorcize

By default, zog starts the server as an attached process. Then the
server is killed when the shell is ended. With the daemonize command,
it's possible to detach the server. The exorcize command is useful
in order to kill the running server.

Note that the server will be killed it you use zog in CLI mode.
The reason is that zog is still not able to check if the server was
started by an other instance or by itself; it's a WIP.
With the Shell mode, you can call server.daemonize in order to be
sure that the server will still running after exit.